Pull up revision 1.3 (requested by tron in ticket #864):   Fix build problem on LP64 platforms.

+++ xfree/xc/extras/Xpm/lib/CrDatFrI.c
@@ -124,7 +124,7 @@ XpmCreateDataFromXpmImage(data_return, image, info)
      */
     header_nlines = 1 + image->ncolors;
     header_size = sizeof(char *) * header_nlines;
-    if (header_size >= SIZE_MAX / sizeof(char *))
+    if (header_size >= UINT_MAX / sizeof(char *))
 	return (XpmNoMemory);
     header = (char **) XpmCalloc(header_size, sizeof(char *));
     if (!header)
